worked on typing

+++ src/data_mod/settings.rs
@@ -9,7 +9,7 @@ pub struct ProgramSettings {
     pub background_color: Color,
 
     pub font_path: String,
-    pub font_size: i64,
+    pub font_size: u32,
     pub font_spacing: f64,
 
     pub cursor_flashing_speed: f64,
@@ -162,7 +162,7 @@ fn get_settings_from_hjson (settings: Map<String, Value>, default_settings: &Pro
         background_color: get_setting_color(&settings, "background color", default_settings.background_color),
 
         font_path: get_setting_lazy(&settings, "font path", |v| v.as_str().map(str::to_string), "String", || default_settings.font_path.to_string()),
-        font_size: get_setting(&settings, "font size", Value::as_i64, "i64", default_settings.font_size),
+        font_size: get_setting(&settings, "font size", Value::as_i64, "i64", default_settings.font_size as i64) as u32,
         font_spacing: get_setting(&settings, "font spacing", Value::as_f64, "f64", default_settings.font_spacing),
 
         cursor_flashing_speed: get_setting(&settings, "cursor flashing speed", Value::as_f64, "f64", default_settings.cursor_flashing_speed),
